Resolved navigation issue

+++ Hotel-backend/Service/Reports/GoalReportService.cs
@@ -91,8 +91,8 @@ public class GoalReportService : IGoalReportService
             M_G = Convert.ToDecimal(goalByMonthGroup == null ? 0 : goalByMonthGroup.ShareRoomM),
             Formatter = "P"
         };
-        int roomSoldByRevenue = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter && x.GroupID == goalArgs.GroupId).SumAsync(x => x.Revenue);
-        int roomSoldByRevenuMonthly = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter).SumAsync(x => x.Revenue);
+        decimal roomSoldByRevenue = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter && x.GroupID == goalArgs.GroupId).SumAsync(x => x.Revenue);
+        decimal roomSoldByRevenuMonthly = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter).SumAsync(x => x.Revenue);
         GoalReportDto marketShareByRevenue = new GoalReportDto
         {
             Indicators = "Market Share based on Revenues",
@@ -103,7 +103,7 @@ public class GoalReportService : IGoalReportService
         };
         var groupIds = classSession.Groups.Select(x => x.GroupId).ToList();
         int revParRoomSold = await _context.RoomAllocation.AsNoTracking().Where(x => x.QuarterNo == quarter && x.GroupID == goalArgs.GroupId).SumAsync(x => x.RoomsAllocated);
-        int groupRoomAllocatedsum = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter && groupIds.Contains(x.GroupID)).SumAsync(x => x.Revenue);
+        decimal groupRoomAllocatedsum = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter && groupIds.Contains(x.GroupID)).SumAsync(x => x.Revenue);
         int roomAllocatedRevPar = 15000 * classSession.Groups.Count;
 
         GoalReportDto RevPar = new GoalReportDto
@@ -116,7 +116,7 @@ public class GoalReportService : IGoalReportService
         };
 
         var adRSoldRoomList = await SoldRoomQueryAsync.Where(x => x.QuarterNo == quarter && groupIds.Contains(x.GroupID)).Select(x => new { x.Revenue, x.SoldRoom }).ToListAsync();
-        int adrGroupRevenuSum = adRSoldRoomList.Sum(x => x.Revenue);
+        decimal adrGroupRevenuSum = adRSoldRoomList.Sum(x => x.Revenue);
         int adrRoomSold = adRSoldRoomList.Sum(x => x.SoldRoom);
         GoalReportDto ADR = new GoalReportDto
         {
@@ -143,7 +143,7 @@ public class GoalReportService : IGoalReportService
 
 
         var incomState = await _context.IncomeState.AsNoTracking().FirstOrDefaultAsync(x => x.QuarterNo == quarter && x.GroupID == goalArgs.GroupId);
-        int roomSoldIncomestate = incomState.Room1 == 0 ? 0 : incomState.Room1 * 100 / 52;
+        decimal roomSoldIncomestate = incomState.Room1 == 0 ? 0 : incomState.Room1 * 100 / 52;
 
         var ScalarMonthAvgIncomeBFcharge = incomeStateMonthlyList.Average(x => x.IncomBfCharg);
         var ScalarMonthAvgRoomRevenue = incomeStateMonthlyList.Average(x => x.Room1);
@@ -160,7 +160,7 @@ public class GoalReportService : IGoalReportService
 
         var ScalarMonthAvgProfit = incomeStateMonthlyList.Average(x => x.NetIncom);
         var ScalarMonthAvgTotalRevenue = incomeStateMonthlyList.Average(x => x.TotReven);
-        int profit = (incomeState.Room1 * 100 / 52);
+        decimal profit = (incomeState.Room1 * 100 / 52);
         GoalReportDto profitMargin = new GoalReportDto
         {
             Indicators = "Profit Margin",
